requirements.txt
fastapi
uvicorn
python-dotenv
PyMuPDF
python-docx
google-generativeai
streamlit
